Re-organises structure, includes .gitignore to never upload samples' effects-lib/ folders
