+++ README.md
@@ -1,2 +1,2 @@
 # corona-sprite-effects
-Controllable / customisable effects for sprite interaction
+Copy the effects-lib/ folder into each sample's root folder to enable it to run.
