+++ effects-lib/effectsdefaults.lua
