+++ samples/README.md
@@ -1,0 +1,2 @@
+# corona-sprite-effects
+Controllable / customisable effects for sprite interaction
