+++ samples/crazy-clock/assets/cog.png
