+++ samples/crazy-clock/assets/corona-icon.png
