+++ samples/crazy-clock/build.settings
