+++ samples/crazy-clock/config.lua
