+++ samples/crazy-clock/libs/globals.lua
