+++ samples/crazy-clock/libs/utils.lua
